orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "shared",
    "a2a-sdk==0.3.26",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.28",
    "pydantic-settings>=2.6",
    "sse-starlette>=2.0",
    "typer>=0.15",
]

[project.scripts]
agentmesh = "orchestrator.cli:main"

[tool.uv.sources]
shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/orchestrator"]
